New test driver for MSCYK, examining hit list.
Additional bugfixes:
- sign error on cutoff conversion to byte
- sign error on setting meta-transition parameters



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@3027 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/impl_sse/sse_cmcons_hitmx.c
@@ -173,7 +173,7 @@ UpdateGammaHitMxCM_epu8(CM_CONSENSUS *ccm, char *errbuf, GammaHitMx_epu8 *gamma,
       assert(jp >= gamma->i0);
       if(do_report_hit) { 
 	/*printf("\t0 %.3f %.3f ip: %d jp: %d r: %d\n", hit_sc+null3_correction, hit_sc, ip, jp, r);*/
-	ReportHit (ip, jp, 0, hit_sc, results);
+	ReportHit (ip, jp, 0, (float) hit_sc, results);
       }
     }
     bestd    = 0;
@@ -191,7 +191,7 @@ UpdateGammaHitMxCM_epu8(CM_CONSENSUS *ccm, char *errbuf, GammaHitMx_epu8 *gamma,
 	  assert(jp >= gamma->i0);
 	  if(do_report_hit) { 
 	    /*printf("\t1 %.3f %.3f ip: %d jp: %d r: %d\n", hit_sc+null3_correction, hit_sc, ip, jp, r);*/
-	    ReportHit (ip, jp, 0, hit_sc, results);
+	    ReportHit (ip, jp, 0, (float) hit_sc, results);
 	  }
 	}
 	if(hit_sc > bestd_sc) { bestd = d; bestd_sc = hit_sc; } /* we need to check again b/c if null3, hit_sc -= null3_correction */
